requirements.txt

streamlit
google-cloud-bigquery
pandas
requests
db-dtypes


